@@ -1,6 +1,6 @@
 {
     "name": "gulp-build-manager",
-    "version": "4.0.2",
+    "version": "4.1.0",
     "description": "Easy to use, flexible gulp task manager",
     "files": [
         "lib",
@@ -10,17 +10,10 @@
     "scripts": {
         "build": "tsc -p src",
         "clean": "rimraf lib",
-        "rebuild": "npm run clean && npm install",
+        "rebuild": "npm run clean && npm run build",
         "watch": "tsc -p src --watch",
-        "docs-build": "gulp docs:@build",
-        "docs-clean": "gulp docs:@clean",
-        "docs-watch": "gulp docs:@watch",
-        "ex-build": "gulp @ex-build-all",
-        "ex-clean": "gulp @ex-clean-all",
-        "build-all-npm-auto": "gulp docs:@build --npm-auto && gulp @ex-build-all --npm-auto",
-        "test": "mocha --reporter spec",
-        "prepare": "npm run clean && npm run build",
-        "prepare-to-commit": "npm run clean && npm run build && npm run test && gulp @clean-all && gulp @build-all && gulp @clean-to-prepare",
+        "test": "mocha test --reporter spec",
+        "prepare-to-commit": "npm run rebuild && npm run test",
         "npm-dry-run": "npm publish --dry-run"
     },
     "repository": {
@@ -49,8 +42,9 @@
     },
     "homepage": "https://github.com/shnam7/gulp-build-manager#readme",
     "dependencies": {
-        "chalk": "^4.0.0",
+        "chalk": "^4.1.0",
         "del": "^5.1.0",
+        "rimraf": "^3.0.2",
         "upath": "^1.2.0"
     },
     "peerDependencies": {
@@ -58,56 +52,9 @@
     },
     "devDependencies": {
         "@types/gulp": "^4.0.6",
-        "@types/gulp-filter": "3.0.33",
-        "@types/jquery": "^3.3.38",
-        "@types/react-dom": "^16.9.7",
-        "autoprefixer": "^9.7.6",
-        "browser-sync": "^2.26.7",
+        "@types/gulp-filter": "^3.0.33",
         "chai": "^4.2.0",
-        "coffeelint-stylish": "^0.1.2",
-        "gulp": "^4.0.2",
-        "gulp-babel": "^8.0.0",
-        "gulp-clean-css": "^4.3.0",
-        "gulp-clone": "^2.0.1",
-        "gulp-coffee": "^3.0.3",
-        "gulp-coffeelint": "^0.6.0",
-        "gulp-concat": "^2.6.1",
-        "gulp-data": "^1.3.1",
-        "gulp-debug": "^4.0.0",
-        "gulp-eslint": "^6.0.0",
-        "gulp-filter": "^6.0.0",
-        "gulp-htmlmin": "^5.0.1",
-        "gulp-imagemin": "^7.1.0",
-        "gulp-less": "^4.0.1",
-        "gulp-markdown": "^5.0.1",
-        "gulp-order": "^1.2.0",
-        "gulp-postcss": "^8.0.0",
-        "gulp-prettier": "^3.0.0",
-        "gulp-rename": "^2.0.0",
-        "gulp-rtlcss": "^1.4.1",
-        "gulp-sass": "^4.1.0",
-        "gulp-sourcemaps": "^2.6.5",
-        "gulp-terser": "^1.2.0",
-        "gulp-twig": "^1.2.0",
-        "gulp-typescript": "^6.0.0-alpha.1",
-        "gulp-zip": "^5.0.1",
-        "lodash.merge": "^4.6.2",
-        "lost": "^8.3.1",
-        "mocha": "^7.1.2",
-        "panini": "^1.6.3",
-        "postcss-combine-duplicated-selectors": "^8.1.0",
-        "postcss-preset-env": "^6.7.0",
-        "postcss-reporter": "^6.0.1",
-        "postcss-utilities": "^0.8.4",
-        "react": "^16.13.1",
-        "react-dom": "^16.13.1",
-        "stream-to-promise": "^2.2.0",
-        "stylelint": "^13.3.3",
-        "stylelint-config-recommended": "^3.0.0",
-        "ts-loader": "^7.0.3",
-        "twig-markdown": "^0.2.1",
-        "typescript": "^3.8.3",
-        "webpack": "^4.43.0",
-        "wicle": "github:shnam7/wicle"
+        "mocha": "^8.1.0",
+        "typescript": "^3.9.7"
     }
 }
