@@ -1,5 +1,5 @@
 describe('gulp-build-manager', function () {
     require('./utils')();
-    require('./buildset')();
     require('./builder')();
+    require('./buildManager')();
 });
